@@ -1,17 +1,17 @@
 name: kraken_animation_player
 description: A kraken plugin which provide `<animation-player />` tags to play animation format files.
-version: 0.1.1
+version: 0.2.0-dev.1
 homepage: https://openkraken.com
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.20.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.2.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flare_flutter: ^2.0.1
-  kraken: ^0.7.0
+  flare_flutter: ^3.0.0
+  kraken: ^0.8.0-dev.1
 
 dev_dependencies:
   flutter_test:
