@@ -1,16 +1,16 @@
 name: kraken_camera
 description: A new flutter plugin project.
-version: 0.1.1
+version: 0.2.0-dev.1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.20.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=2.2.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  kraken: ^0.7.1
-  path_provider: ^1.1.0
+  kraken: ^0.8.0-dev.1
+  path_provider: ^2.0.2
 
 dev_dependencies:
   flutter_test:
