@@ -18,7 +18,7 @@ import 'camera_mixins.dart';
 ///
 /// The [textureId] can be passed to a [Texture] widget.
 class NativeTexture with CameraMappable {
-  NativeTexture._({@required int handle, @required this.textureId})
+  NativeTexture._({required int handle, required this.textureId})
       : _handle = handle,
         assert(handle != null),
         assert(textureId != null);
@@ -33,10 +33,10 @@ class NativeTexture with CameraMappable {
   static Future<NativeTexture> allocate() async {
     final int handle = CameraChannel.nextHandle++;
 
-    final int textureId = await CameraChannel.channel.invokeMethod<int>(
+    final int textureId = await (CameraChannel.channel.invokeMethod<int>(
       '$NativeTexture#allocate',
       <String, dynamic>{'textureHandle': handle},
-    );
+    ) as FutureOr<int>);
 
     return NativeTexture._(handle: handle, textureId: textureId);
   }
