@@ -7,14 +7,14 @@ import Foundation
 
 import connectivity_macos
 import kraken
+import kraken_animation_player
 import path_provider_macos
 import shared_preferences_macos
-import kraken_animation_player
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   ConnectivityPlugin.register(with: registry.registrar(forPlugin: "ConnectivityPlugin"))
   KrakenPlugin.register(with: registry.registrar(forPlugin: "KrakenPlugin"))
+  KrakenAnimationPlayerPlugin.register(with: registry.registrar(forPlugin: "KrakenAnimationPlayerPlugin"))
   PathProviderPlugin.register(with: registry.registrar(forPlugin: "PathProviderPlugin"))
   SharedPreferencesPlugin.register(with: registry.registrar(forPlugin: "SharedPreferencesPlugin"))
-  KrakenAnimationPlayerPlugin.register(with: registry.registrar(forPlugin: "KrakenAnimationPlayerPlugin"))
 }
