@@ -25,7 +25,7 @@ enum LensDirection { front, back, unknown }
 /// the lens face.
 abstract class CameraDescription {
   /// Location of the camera on the device.
-  LensDirection get direction;
+  LensDirection? get direction;
 
   /// Identifier for this camera.
   String get name;
