@@ -23,7 +23,7 @@ class Camera with NativeMethodCallHandler {
   bool _isClosed = false;
 
   /// Retrieves the number of physical cameras available on this device.
-  static Future<int> getNumberOfCameras() {
+  static Future<int?> getNumberOfCameras() {
     return CameraChannel.channel.invokeMethod<int>(
       'Camera#getNumberOfCameras',
     );
@@ -55,10 +55,10 @@ class Camera with NativeMethodCallHandler {
   /// If [getNumberOfCameras] returns N, the valid id is 0 to N-1.
   static Future<CameraInfo> getCameraInfo(int cameraId) async {
     final Map<String, dynamic> infoMap =
-        await CameraChannel.channel.invokeMapMethod<String, dynamic>(
+        await (CameraChannel.channel.invokeMapMethod<String, dynamic>(
       'Camera#getCameraInfo',
       <String, dynamic>{'cameraId': cameraId},
-    );
+    ) as FutureOr<Map<String, dynamic>>);
 
     return CameraInfo.fromMap(infoMap);
   }
