@@ -7,14 +7,14 @@ import Foundation
 
 import connectivity_macos
 import kraken
+import kraken_mqtt
 import path_provider_macos
 import shared_preferences_macos
-import kraken_mqtt
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   ConnectivityPlugin.register(with: registry.registrar(forPlugin: "ConnectivityPlugin"))
   KrakenPlugin.register(with: registry.registrar(forPlugin: "KrakenPlugin"))
+  KrakenMqttPlugin.register(with: registry.registrar(forPlugin: "KrakenMqttPlugin"))
   PathProviderPlugin.register(with: registry.registrar(forPlugin: "PathProviderPlugin"))
   SharedPreferencesPlugin.register(with: registry.registrar(forPlugin: "SharedPreferencesPlugin"))
-  KrakenMqttPlugin.register(with: registry.registrar(forPlugin: "KrakenMqttPlugin"))
 }
