@@ -7,14 +7,14 @@ import Foundation
 
 import connectivity_macos
 import kraken
+import kraken_websocket
 import path_provider_macos
 import shared_preferences_macos
-import kraken_websocket
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   ConnectivityPlugin.register(with: registry.registrar(forPlugin: "ConnectivityPlugin"))
   KrakenPlugin.register(with: registry.registrar(forPlugin: "KrakenPlugin"))
+  KrakenWebsocketPlugin.register(with: registry.registrar(forPlugin: "KrakenWebsocketPlugin"))
   PathProviderPlugin.register(with: registry.registrar(forPlugin: "PathProviderPlugin"))
   SharedPreferencesPlugin.register(with: registry.registrar(forPlugin: "SharedPreferencesPlugin"))
-  KrakenWebsocketPlugin.register(with: registry.registrar(forPlugin: "KrakenWebsocketPlugin"))
 }
