@@ -7,14 +7,14 @@ import Foundation
 
 import connectivity_macos
 import kraken
+import kraken_webview
 import path_provider_macos
 import shared_preferences_macos
-import kraken_webview
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   ConnectivityPlugin.register(with: registry.registrar(forPlugin: "ConnectivityPlugin"))
   KrakenPlugin.register(with: registry.registrar(forPlugin: "KrakenPlugin"))
+  KrakenWebviewPlugin.register(with: registry.registrar(forPlugin: "KrakenWebviewPlugin"))
   PathProviderPlugin.register(with: registry.registrar(forPlugin: "PathProviderPlugin"))
   SharedPreferencesPlugin.register(with: registry.registrar(forPlugin: "SharedPreferencesPlugin"))
-  KrakenWebviewPlugin.register(with: registry.registrar(forPlugin: "KrakenWebviewPlugin"))
 }
